wip for characters
